split up host
