+++ guest/hat/src/lib.rs
@@ -95,7 +95,7 @@ fn read_temperature(i2c: &mut I2c) -> i16 {
     }
 }
 impl Guest for Component {
-    fn run(mut connection: I2c) -> String {
+    fn get(mut connection: I2c) -> String {
         while !data_available(&mut connection) {}
 
         let temperature_x8 = read_temperature(&mut connection);
