+++ host/.cargo/config.toml
@@ -1,2 +1,5 @@
+[build]
+target = "aarch64-unknown-linux-gnu"
+
 [target.aarch64-unknown-linux-gnu]
 linker = "aarch64-linux-gnu-gcc"
