--- wit/screen.wit
@@ -1,15 +1,0 @@
-package sketch:embedded@0.0.0;
-
-interface run {
-    use delay.{delay};
-    use i2c.{i2c};
-
-    run: func(connection: i2c, delay: delay);
-}
-
-world screen {
-    import delay;
-    import i2c;
-
-    export run;
-}
