--- wit/sensor.wit
@@ -1,13 +1,0 @@
-package sketch:embedded@0.0.0;
-
-interface temperature {
-    use i2c.{i2c};
-
-    run: func(connection: i2c) -> string;
-}
-
-world sensor {
-    import i2c;
-
-    export temperature;
-}
